peavy_demo/requirements.txt
django-peavy
psycopg2==2.4.1   # 2.4.2 causes trouble with Django 1.3(.1) unit tests
